requirements.txt
flask==2.0.1
werkzeug==2.0.1
python-dotenv==0.19.0
requests==2.26.0
deepl==1.11.0
spotipy==2.23.0 